+++ src/app/app.server.module.ts
@@ -5,6 +5,22 @@ import { IonicServerModule } from '@ionic/angular-server';
 
 import { AppModule } from './app.module';
 import { AppComponent } from './app.component';
+import {
+  RouteStateSourceService,
+  ServerRouteStateSourceService
+} from '@services/route-state-source.service';
+import {
+  CollectionTextViewsQueryParamSyncService,
+  ServerCollectionTextViewsQueryParamSyncService
+} from '@services/collection-text-views-query-param-sync.service';
+import {
+  RouterNavigationSourceService,
+  ServerRouterNavigationSourceService
+} from '@services/router-navigation-source.service';
+import {
+  RouterPreloadingStrategyService,
+  ServerRouterPreloadingStrategyService
+} from '@services/router-preloading-strategy.service';
 
 
 @NgModule({
@@ -13,7 +29,25 @@ import { AppComponent } from './app.component';
     ServerModule,
     IonicServerModule,
   ],
-  providers: [provideHttpClient(withFetch())],
+  providers: [
+    provideHttpClient(withFetch()),
+    {
+      provide: RouteStateSourceService,
+      useClass: ServerRouteStateSourceService
+    },
+    {
+      provide: CollectionTextViewsQueryParamSyncService,
+      useClass: ServerCollectionTextViewsQueryParamSyncService
+    },
+    {
+      provide: RouterNavigationSourceService,
+      useClass: ServerRouterNavigationSourceService
+    },
+    {
+      provide: RouterPreloadingStrategyService,
+      useClass: ServerRouterPreloadingStrategyService
+    }
+  ],
   bootstrap: [AppComponent],
 })
 export class AppServerModule {}